refactor(ui): extract BaseCard elements into independent components
