+++ app/shared/components/base-card/TitleWithTooltip.style.ts
@@ -1,0 +1,16 @@
+import { SxProps, Theme } from '@mui/material';
+
+const styles: Record<string, any> = {
+  title: (lineClamp: number = 2, fontWeight: number = 700): SxProps<Theme> => ({
+    fontWeight: fontWeight,
+    color: 'text.primary',
+    flex: 1,
+    minWidth: 0,
+    display: '-webkit-box',
+    WebkitLineClamp: lineClamp,
+    WebkitBoxOrient: 'vertical',
+    overflow: 'hidden',
+  }),
+};
+
+export default styles;
